move old host table and container next to other code related to host page
