+++ frontend/pages/hosts/ManageHostsPage/components/HostContainer/_styles.scss
