+++ frontend/pages/hosts/ManageHostsPage/components/HostContainer/index.js
