+++ frontend/pages/hosts/ManageHostsPage/components/HostsTable/_styles.scss
