+++ frontend/pages/hosts/ManageHostsPage/components/HostsTable/index.js
